examples/efm32/r1k1: fix the stupid type issue

+++ examples/efm32/r1k1/main.cpp
@@ -24,7 +24,7 @@ void loop() {
 	Serial.println(a7);
 
 	Serial.print("Rt = ");
-	Serial.println( 1100.0 / ((float)a6 / a7 - 1.0), 1 );
+	Serial.println( 1100.0 / ((float)a6 / (float)a7 - 1.0), 1 );
 
 	delay(2000);  // wait for 2 seconds
 }
